添加有状态的部件，设置TabBarView项
